feat(productlistings): move to proper folder
[CAPI-85]
